devops py formatting and test placeholders
