+++ src/FunctionApps/DevOps/.vscode/settings.json
@@ -2,5 +2,10 @@
   "azureFunctions.deploySubpath": ".",
   "azureFunctions.scmDoBuildDuringDeployment": true,
   "azureFunctions.pythonVenv": ".venv",
-  "azureFunctions.projectLanguage": "Python"
+  "azureFunctions.projectLanguage": "Python",
+  "python.testing.pytestArgs": [
+    "tests"
+  ],
+  "python.testing.unittestEnabled": false,
+  "python.testing.pytestEnabled": true
 }
