+++ src/FunctionApps/DevOps/requirements_dev.txt
@@ -1,1 +1,7 @@
+# DO NOT include azure-functions-worker in this file
+# The Python Worker is managed by Azure Functions platform
+# Manually managing azure-functions-worker may cause unexpected issues
+
+mypy
 pytest
+black
