+++ src/FunctionApps/DevOps/tests/test_confirm_storage_access.py
@@ -1,0 +1,3 @@
+def test_confirm_storage_access_placeholder():
+    """placeholder so pytest does not fail"""
+    pass
