+++ src/FunctionApps/DevOps/tests/test_get_ip.py
@@ -1,0 +1,3 @@
+def test_get_ip_placeholder():
+    """placeholder so pytest does not fail"""
+    pass
